move publicPathPrefix.ts to @lib/utils
