+++ ui/lib/apps/Diagnose/pages/DiagnoseStatus.tsx
@@ -5,7 +5,7 @@ import { Link, useParams } from 'react-router-dom'
 import { ArrowLeftOutlined } from '@ant-design/icons'
 
 import client from '@lib/client'
-import publicPathPrefix from '@dashboard/publicPathPrefix'
+import publicPathPrefix from '@lib/utils/publicPathPrefix'
 import { AnimatedSkeleton, DateTime, Head } from '@lib/components'
 import { useClientRequestWithPolling } from '@lib/utils/useClientRequest'
