+++ ui/public/dbassPortal.html
@@ -6,7 +6,7 @@
 
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="description" content="TiDB Dashboard" />
+    <meta name="description" content="TiDB Dashboard DBass Portal" />
     <!-- prettier-ignore -->
     <style>body{margin:0;font-family:-apple-system,BlinkMacSystemFont,'Segoe UI',Roboto,Oxygen,Ubuntu,Cantarell,'Fira Sans','Droid Sans','Helvetica Neue',sans-serif;-webkit-font-smoothing:antialiased;-moz-osx-font-smoothing:grayscale;background:#f0f2f5}#dashboard_page_spinner{font-size:10px;text-indent:-9999em;width:8em;height:8em;position:fixed;margin-left:-4em;margin-top:-4em;left:50%;top:50%;border-radius:50%;background:#a1d8f8;background:-moz-linear-gradient(left,#a1d8f8 10%,rgba(161,216,248,0) 42%);background:-webkit-linear-gradient(left,#a1d8f8 10%,rgba(161,216,248,0) 42%);background:-o-linear-gradient(left,#a1d8f8 10%,rgba(161,216,248,0) 42%);background:-ms-linear-gradient(left,#a1d8f8 10%,rgba(161,216,248,0) 42%);background:linear-gradient(to right,#a1d8f8 10%,rgba(161,216,248,0) 42%);-webkit-animation:load3 1.4s infinite linear;animation:load3 1.4s infinite linear;-webkit-transform:translateZ(0);-ms-transform:translateZ(0);transform:translateZ(0)}#dashboard_page_spinner:before{width:50%;height:50%;background:#a1d8f8;border-radius:100% 0 0 0;position:absolute;top:0;left:0;content:''}#dashboard_page_spinner:after{background:#f0f2f5;width:75%;height:75%;border-radius:50%;content:'';margin:auto;position:absolute;top:0;left:0;bottom:0;right:0}@-webkit-keyframes load3{0%{-webkit-transform:rotate(0);transform:rotate(0)}100%{-webkit-transform:rotate(360deg);transform:rotate(360deg)}}@keyframes load3{0%{-webkit-transform:rotate(0);transform:rotate(0)}100%{-webkit-transform:rotate(360deg);transform:rotate(360deg)}}</style>
     <title>TiDB Dashboard</title>
