+++ ui/src/publicPath.ts
@@ -1,2 +1,2 @@
-import publicPathPrefix from '@dashboard/publicPathPrefix'
+import publicPathPrefix from '@lib/utils/publicPathPrefix'
 __webpack_public_path__ = publicPathPrefix + '/'
